Rename array->exact
